-Created Group validation service -Refactored code and method calls accordingly
